extension/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.10.0
python-multipart>=0.0.9
httpx>=0.27.0
pillow>=11.0.0
numpy>=2.0.0
beautifulsoup4>=4.12.3
python-dotenv>=1.0.0